requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
openai
python-multipart
google-generativeai
